Refactored GameScreen Everything related to tilemap moved into Board class. Moved update and canGo from GameScreen to Player. Minor changes in Player. Player tests are not working (issue related to Board).
