--- src/main/java/inf112/skeleton/app/Direction.java
@@ -1,24 +1,0 @@
-package inf112.skeleton.app;
-
-
-public enum Direction {
-
-    NORTH,
-    EAST,
-    SOUTH,
-    WEST;
-
-    private Direction opposite;
-
-    static {
-        NORTH.opposite = SOUTH;
-        SOUTH.opposite = NORTH;
-        EAST.opposite = WEST;
-        WEST.opposite = EAST;
-    }
-
-
-    public Direction getOppositeDirection() {
-        return opposite;
-    }
-}
