+++ src/main/java/inf112/skeleton/app/GameWorld/Direction.java
@@ -1,0 +1,24 @@
+package inf112.skeleton.app.GameWorld;
+
+
+public enum Direction {
+
+    NORTH,
+    EAST,
+    SOUTH,
+    WEST;
+
+    private Direction opposite;
+
+    static {
+        NORTH.opposite = SOUTH;
+        SOUTH.opposite = NORTH;
+        EAST.opposite = WEST;
+        WEST.opposite = EAST;
+    }
+
+
+    public Direction getOppositeDirection() {
+        return opposite;
+    }
+}
