+++ src/main/java/inf112/skeleton/app/Rotate.java
@@ -1,7 +1,6 @@
 package inf112.skeleton.app;
 
 
-//TODO? Rename
 public enum Rotate {
     NONE,
     LEFT,
