+++ src/test/java/inf112/skeleton/app/MovePlayerTest.java
@@ -1,6 +1,9 @@
 package inf112.skeleton.app;
 
 
+import inf112.skeleton.app.GameObjects.Player;
+import inf112.skeleton.app.GameWorld.Board;
+import inf112.skeleton.app.GameWorld.Direction;
 import org.junit.Before;
 import org.junit.Test;
 
@@ -14,7 +17,8 @@ public class MovePlayerTest {
 
     @Before
     public void initialize() {
-        player = new Player("Player1", x, y, Direction.SOUTH);
+        Board board = new Board(Main.VAULT);
+        player = new Player("Player1", x, y, Direction.SOUTH, board);
 
     }
