+++ src/test/java/inf112/skeleton/app/PlayerTest.java
@@ -1,5 +1,8 @@
 package inf112.skeleton.app;
 
+import inf112.skeleton.app.GameObjects.Player;
+import inf112.skeleton.app.GameWorld.Board;
+import inf112.skeleton.app.GameWorld.Direction;
 import org.junit.Before;
 import org.junit.Test;
 
@@ -16,7 +19,8 @@ public class PlayerTest {
 
     @Before
     public void setup(){
-        player = new Player("testBot", 0,0, Direction.SOUTH);
+        Board board = new Board(Main.VAULT);
+        player = new Player("testBot", 0,0, Direction.SOUTH, board);
         stack = ProgramCard.makeStack();
     }
 
@@ -172,9 +176,10 @@ public class PlayerTest {
     public void priorityTest(){
         PriorityQueue<ProgramCard> q = new PriorityQueue<>();
         // create 3 players:
-        Player p1 = new Player("p1", 0,0);
-        Player p2 = new Player("p2", 0,1);
-        Player p3 = new Player("p3", 0,2);
+        Board board = new Board(Main.VAULT);
+        Player p1 = new Player("p1", 0,0, Direction.SOUTH, board);
+        Player p2 = new Player("p2", 0,1, Direction.SOUTH, board);
+        Player p3 = new Player("p3", 0,2, Direction.SOUTH, board);
         // give them five cards each:
         for(int i = 0; i < 5; i++){
             p1.receiveNewCard(stack.pop());
